fix(activity-thread): refactor to not use context and refs
